requirements.txt
fastapi==0.115.12
langchain==0.3.24
langchain_community==0.3.23
langchain_core==0.3.57
langchain_groq==0.3.2
langchain_openai==0.3.15
langchain_pinecone==0.2.5
langchain_text_splitters==0.3.8
langgraph==0.4.1
mistralai==1.7.0
openai==1.76.2
passlib==1.7.4
pinecone==6.0.2
pydantic==2.11.4
pydantic_settings==2.9.1
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
